Moving the exposing of the load function to App.vue This is to make it available even in the ladder. Needing to switch once to options menu is a bit weird.
